Looking for api data and showing on screen
